+++ app/(navigatable)/(projects)/projects/project-details-tile.tsx
@@ -68,7 +68,7 @@ export const ProjectDetailsTile = ({
       href={`/projects/${id}`}
       className="block transform transition hover:scale-105"
     >
-      <div className="flex justify-between bg-background shadow-md rounded-lg p-6 hover:bg-blue-500">
+      <div className="flex justify-between bg-background shadow-md rounded-lg border border-blue-500 p-6 hover:bg-blue-500">
         <div>
           <h3 className="text-lg font-semibold">{name}</h3>
           <p className="text-sm text-gray-600">